1/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
pypdf>=3.17.0
pydantic>=2.0.0
